# docs/openapi.implemented.yaml
...
components:
  schemas:
...
    AttributeScoreBody:
      properties:
        attribute:
          title: Attribute
          type: string
        contribution:
          title: Contribution
          type: number
        effective_weight:
          title: Effective Weight
          type: number
        normalised_score:
          anyOf:
          - type: integer
          - type: 'null'
          title: Normalised Score
        pillar:
          title: Pillar
          type: string
      required:
      - attribute
      - pillar
      - effective_weight
      - contribution
      title: AttributeScoreBody
      type: object
...
    CandidateResultBody:
      properties:
        candidate:
          title: Candidate
          type: string
        coverage:
          title: Coverage
          type: number
        coverage_by_confidence:
          anyOf:
          - $ref: '#/components/schemas/ConfidenceSplitBody'
          - type: 'null'
          description: How the covered weight splits by confidence. Null when nothing
            is covered.
        insufficient_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Insufficient Reason
        match_status:
          title: Match Status
          type: string
        name:
          title: Name
          type: string
        rank:
          anyOf:
          - type: integer
          - type: 'null'
          title: Rank
        score:
          anyOf:
          - type: integer
          - type: 'null'
          description: Null only when insufficient data. On the evaluation's scale.
          title: Score
      required:
      - candidate
      - name
      - coverage
      - match_status
      title: CandidateResultBody
      type: object
    CandidateScoreDetailBody:
      properties:
        attribute_scores:
          default: []
          items:
            $ref: '#/components/schemas/AttributeScoreBody'
          title: Attribute Scores
          type: array
        candidate:
          title: Candidate
          type: string
        coverage:
          title: Coverage
          type: number
        coverage_by_confidence:
          anyOf:
          - $ref: '#/components/schemas/ConfidenceSplitBody'
          - type: 'null'
          description: How the covered weight splits by confidence. Null when nothing
            is covered.
        insufficient_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Insufficient Reason
        match_status:
          title: Match Status
          type: string
        name:
          title: Name
          type: string
        rank:
          anyOf:
          - type: integer
          - type: 'null'
          title: Rank
        score:
          anyOf:
          - type: integer
          - type: 'null'
          description: Null only when insufficient data. On the evaluation's scale.
          title: Score
      required:
      - candidate
      - name
      - coverage
      - match_status
      title: CandidateScoreDetailBody
      type: object
...
    ConfidenceSplitBody:
      description: Percentages of the covered weight, summing to 100 (`reqs.md` 5.7).
      properties:
        absolute:
          title: Absolute
          type: number
        high:
          title: High
          type: number
        low:
          title: Low
          type: number
        medium:
          title: Medium
          type: number
      required:
      - absolute
      - high
      - medium
      - low
      title: ConfidenceSplitBody
      type: object
...
    CriterionBody:
      additionalProperties: true
      type: object
...
    EvaluationCriteriaBody:
      properties:
        criteria:
          items:
            $ref: '#/components/schemas/CriterionBody'
          title: Criteria
          type: array
        pillar_weights:
          items:
            $ref: '#/components/schemas/EvaluationPillarWeightBody'
          title: Pillar Weights
          type: array
      required:
      - pillar_weights
      - criteria
      title: EvaluationCriteriaBody
      type: object
    EvaluationPillarWeightBody:
      description: 'The contract''s `PillarWeight`, which names no level: the evaluation''s
        header already

        says which level it ran at, and an evaluation never spans two (`reqs.md` 3.4a).'
      properties:
        pillar:
          title: Pillar
          type: string
        weight:
          title: Weight
          type: number
        weight_locked:
          title: Weight Locked
          type: boolean
      required:
      - pillar
      - weight
      - weight_locked
      title: EvaluationPillarWeightBody
      type: object
    EvaluationSummaryBody:
      properties:
        computed_at:
          format: date-time
          title: Computed At
          type: string
        criteria_set:
          title: Criteria Set
          type: string
        id:
          title: Id
          type: integer
        level:
          title: Level
          type: string
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        score_scale_max:
          title: Score Scale Max
          type: integer
      required:
      - id
      - criteria_set
      - level
      - computed_at
      - score_scale_max
      title: EvaluationSummaryBody
      type: object
    EvaluationsBody:
      properties:
        items:
          items:
            $ref: '#/components/schemas/EvaluationSummaryBody'
          title: Items
          type: array
      required:
      - items
      title: EvaluationsBody
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    RankingBody:
      properties:
        candidates:
          items:
            $ref: '#/components/schemas/CandidateResultBody'
          title: Candidates
          type: array
        computed_at:
          format: date-time
          title: Computed At
          type: string
        criteria_set:
          title: Criteria Set
          type: string
        level:
          title: Level
          type: string
      required:
      - criteria_set
      - level
      - computed_at
      - candidates
      title: RankingBody
      type: object
...
    SaveEvaluationBody:
      properties:
        criteria_set:
          title: Criteria Set
          type: string
        level:
          title: Level
          type: string
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      required:
      - criteria_set
      - level
      title: SaveEvaluationBody
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
info:
  title: Starnest
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /v1/evaluations:
    get:
      description: The kept results, newest first.
      operationId: listEvaluations
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationsBody'
          description: Successful Response
      summary: List Evaluations
      tags:
      - evaluation
    post:
      description: 'Compute the ranking and keep it, with the criteria snapshot behind
        it.


        The same computation `GET /rankings` performs, called rather than repeated:
        two code paths

        would let a saved evaluation differ from the ranking it was saved from.'
      operationId: saveEvaluation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveEvaluationBody'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationSummaryBody'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Save Evaluation
      tags:
      - evaluation
  /v1/evaluations/{evaluation_id}:
    get:
      description: A kept ranking, in the order it was kept. Raises a 404 for an evaluation
        nobody kept.
      operationId: getEvaluation
      parameters:
      - in: path
        name: evaluation_id
        required: true
        schema:
          title: Evaluation Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RankingBody'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Evaluation
      tags:
      - evaluation
  /v1/evaluations/{evaluation_id}/candidates/{candidate_id}:
    get:
      description: Why this candidate scored what it scored, attribute by attribute.
      operationId: getCandidateScoreDetail
      parameters:
      - in: path
        name: evaluation_id
        required: true
        schema:
          title: Evaluation Id
          type: integer
      - in: path
        name: candidate_id
        required: true
        schema:
          title: Candidate Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateScoreDetailBody'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Candidate Score Detail
      tags:
      - evaluation
  /v1/evaluations/{evaluation_id}/criteria:
    get:
      description: 'The frozen snapshot: what the weights, goals, bands and anchors
        actually were.'
      operationId: getEvaluationCriteria
      parameters:
      - in: path
        name: evaluation_id
        required: true
        schema:
          title: Evaluation Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationCriteriaBody'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Evaluation Criteria
      tags:
      - evaluation
...
  /v1/rankings:
    get:
      description: 'Score every candidate at this level and return them in rank order.


        **Computes and returns; stores nothing** (`reqs.md` 5.6, Q155). Adjusting
        a weight

        recalculates from stored values, and an afternoon of tuning must not bury
        the few results

        worth keeping under hundreds nobody asked for. `POST /evaluations` is how
        one is kept.'
      operationId: getRanking
      parameters:
      - in: query
        name: criteria_set
        required: true
        schema:
          title: Criteria Set
          type: string
      - in: query
        name: level
        required: true
        schema:
          title: Level
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RankingBody'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Ranking
      tags:
      - rankings
...